simplify revision data size handling - just store it into metadata
Back when the storage api was implemented, size was implemented as a property.
I can't remember why this was done, but I assume it was to do backend-specific
optimization to quickly access revision data size without accessing metadata.

As we'll have an index for most important metadata and we likely have to open
metadata anyway, SIZE is now just a metadata entry like all the others, not
specialcased any more.

Like the revision data hash, SIZE is also automatically stored into the
revision's metadata when committing it.

This simplifies the code quite a bit and makes it more generic.