make converter api more homogeneous, adapt tests
you can give a file-like object (like a revision), a str or unicode now to a
input converter (no matter for what mimetype it is).

you also need to give the contenttype (at least for the text input converters,
they will extract the charset from there and decode appropriately).

removed the feed_input_conv methods, we always give the revision and the
contenttype now to the converters.