use converters to convert a rev to indexable content
revision data may contain all sorts of content-types, e.g.:
- markup text (moin, creole, docbook, rst, html, ...)
- plain text (code, docs, ...)
- images, audio
- archives and other binary stuff

for rendering this, we already use the (dom) converters.

for indexing this, we can also use them, this is a proof of concept implementation
for this. it shows some problems still (likely due to bugs in other converters),
but for some types it works.

use the /+indexable/foo view to see what it makes out of foo's content.

currently it outputs rst markup (because we already have an output converter for
this), but this should be changed to a simple text output converter that outputs
no markup, just text.