history: do not yield Revision objects, but directly result documents from whoosh
we have all that global and local history view (and also the atom feed) needs
in the whoosh index. the whoosh documents yielded are dicts with all stored fields.

by just using that, we do not need to access the backend storage any more for
displaying history (which was one of the reasons for the slow global history
for wikis with many revisions).

one small difference is rev.timestamp (UNIX timestamp) vs. doc[MTIME] (datetime
object).

added size, action, comment fields to whoosh indexes (needed for history
display and also useful for other reasons).

adjusted the tests (pytest 2)

TODO: ACL checks for history? (this is a general thing we need to be careful
with: not to expose index data that should not be exposed).

removed unused "mountpoint" history() param.

add utctimestamp() == inverse of datetime.utcfromtimestamp()