fix cache key computation bug
we used the hash of the data as cache key at some places.
but as we also have metadata and metadata might be different for same data hash,
this leads to problems. solved now by just using the revid (which is a UUID).

this fixes some tests that were broken here since enabling usage of app.cache,
as often data was just '' (-> always same hash), but contenttype was sometimes
a wiki page, sometimes an image, so e.g. the internal representation was
expected to be different. As it used a wrong cached internal representation,
tests were failing...